requirements.txt
pillow >=7, <8
